+++ src/tools/ahkClass/getClassPropertyDef.ts
@@ -2,16 +2,28 @@ import * as vscode from 'vscode';
 import type { CAhkClassPropertyGetSet } from '../../AhkSymbol/CAhkClass';
 import { CAhkClassPropertyDef } from '../../AhkSymbol/CAhkClass';
 import { getChildren, type TFuncInput } from '../../core/getChildren';
+import type { TTokenStream } from '../../globalEnum';
 import { getRange } from '../range/getRange';
 import { getCAhkClassGet } from './getCAhkClassGet';
+
+function getPropertySearchLine(DocStrMap: TTokenStream, defLine: number): number | null {
+    for (let i: number = defLine + 1; i < DocStrMap.length; i++) {
+        const { cll, lStr, line } = DocStrMap[i];
+        if ((/^[ \t]*\{/u).test(lStr)) return line;
+        if (cll === 0) {
+            return null;
+        }
+    }
+    return null;
+}
 
 export function getClassPropertyDef(FuncInput: TFuncInput): CAhkClassPropertyDef | null {
     const { lStr, line } = FuncInput.AhkTokenLine;
 
     if (lStr.includes('(') || lStr.includes('=')) return null;
 
-    const ma: RegExpMatchArray | null = lStr.match(/^[ \t]*(?:\}[ \t]*)?([#$@\w\u{A1}-\u{FFFF}]+)[ \t]*\[/u)
-        ?? lStr.match(/^[ \t]*(?:\}[ \t]*)?([#$@\w\u{A1}-\u{FFFF}]+)(?:[ \t{]|$)/u);
+    const ma: RegExpMatchArray | null = lStr.match(/^[ \t}]*([#$@\w\u{A1}-\u{FFFF}]+)[ \t]*\[/u)
+        ?? lStr.match(/^[ \t}]*([#$@\w\u{A1}-\u{FFFF}]+)(?:[ \t{]|$)/u);
     if (ma === null) return null;
 
     const {
@@ -21,18 +33,25 @@ export function getClassPropertyDef(FuncInput: TFuncInput): CAhkClassPropertyDef
         GValMap,
     } = FuncInput;
 
-    if (ma[0].endsWith('[')) {
-        // TODO something ...
-    }
+    // if (ma[0].endsWith('[')) {
+    //     // TODO something ...
+    // }
 
     const name: string = ma[1];
-    const col: number = lStr.indexOf(name);
+    const colS: number = lStr.indexOf(name);
+    const colE: number = colS + name.length;
     const selectionRange: vscode.Range = new vscode.Range(
-        new vscode.Position(line, col),
-        new vscode.Position(line, col + name.length),
+        new vscode.Position(line, colS),
+        new vscode.Position(line, colE),
     );
 
-    const range = getRange(DocStrMap, line, line, RangeEndLine, col);
+    const searchLine: number | null = lStr.includes('{')
+        ? line
+        : getPropertySearchLine(DocStrMap, line);
+
+    if (searchLine === null) return null;
+
+    const range = getRange(DocStrMap, line, searchLine, RangeEndLine, colS);
 
     const ch: CAhkClassPropertyGetSet[] = getChildren<CAhkClassPropertyDef>(
         [getCAhkClassGet],